Fix remaining float-to-int deprecations found via a repo-wide audit
Per operator request: swept the whole ui/ tree with PHPStan level 5
(--error-format=raw, filtered for "float given" against a typed int
parameter) to make sure the graph-rendering OOM bug's underlying
deprecation class ("Implicit conversion from float to int loses
precision", PHP 8.1) had no other live instances, not just the ones
already hit by manual testing.

Found and fixed:
- draw.inc.php: zbx_imageline()'s and zbx_imagealine()'s own round()/
  floor() calls return float in PHP — every coordinate they compute was
  still being handed to imageline()/imagesetpixel()/imagecolorat() as a
  float despite the wrapper's whole purpose being "give GD an integer
  coordinate". zbx_imagealine() runs its float math inside a per-pixel
  loop for every plain LINE/BOLD_LINE-drawtype graph (not just the
  gradient/filled-region drawtype fixed previously), so this was a
  second, wider path to the same failure mode.
- CLineGraphDraw.php: the non-working-hours background shading loop
  (round()/ceil() results passed straight to imagefilledrectangle())
  triggers once per work/non-work period boundary in the requested time
  range — negligible for a 1-hour graph but scales with the range.

Verified the remaining PHPStan "float given" hits outside these files
(func.inc.php strlen(), items.inc.php CHistoryManager::getAggregatedValue(),
CWidgetForm.php CWidgetFieldIntegerBox) are PHPDoc-only annotations with
no native type hint on the receiving parameter — PHP performs no type
coercion there at all, so none of them can trigger this deprecation.
No change needed for those.

Re-verified all 19 previously-sampled graphs over a 7-day range
(exercises the work-period shading loop far more than the original 1h
range) render cleanly with zero deprecation messages logged. PHPUnit
21/21 still green.

+++ sources/zabbix-5.0.47/ui/include/classes/graphdraw/CLineGraphDraw.php
@@ -1316,8 +1316,8 @@ class CLineGraphDraw extends CGraphDraw {
 		while ($start < $max_time && $start > 0) {
 			$end = find_period_end($periods, $start, $max_time);
 
-			$x1 = round((($start - $from) * $this->sizeX) / $this->period) + $this->shiftXleft;
-			$x2 = ceil((($end - $from) * $this->sizeX) / $this->period) + $this->shiftXleft;
+			$x1 = (int) round((($start - $from) * $this->sizeX) / $this->period) + $this->shiftXleft;
+			$x2 = (int) ceil((($end - $from) * $this->sizeX) / $this->period) + $this->shiftXleft;
 
 			// draw rectangle
 			imagefilledrectangle(
